znisczylem sklep w tym commicie

+++ src/code/Produkt/Obuwie.java
@@ -8,7 +8,7 @@ public class Obuwie extends Produkt implements Serializable {
     private String typObuwia;
     private String typPodeszwy;
 
-    public Obuwie(String idProduktu, double cena, String nazwa, TreeMap rozmiary, String opis, String material,
+    public Obuwie(String idProduktu, float cena, String nazwa, TreeMap rozmiary, String opis, String material,
                   String kolor, Producent producent, String typObuwia, String typPodeszwy, ImageIcon icon){
         super(idProduktu, cena, nazwa, rozmiary, opis, material, kolor, producent, icon);
         this.typObuwia=typObuwia;
